feat: Implement coordinator logic, use uint64_t, add new version of StateCheckRpc

+++ src/coordination/raft_state.cpp
@@ -434,7 +434,7 @@ auto RaftState::GetEnabledReadsOnMain() const -> bool { return state_machine_->G
 
 auto RaftState::GetSyncFailoverOnly() const -> bool { return state_machine_->GetSyncFailoverOnly(); }
 
-auto RaftState::GetMaxReplicaLag() const -> uint32_t { return state_machine_->GetMaxReplicaLag(); }
+auto RaftState::GetMaxReplicaLag() const -> uint64_t { return state_machine_->GetMaxReplicaLag(); }
 
 }  // namespace memgraph::coordination
